Change integer types for size and index fields to i64/i16
Updated the types of 'size' and 'index' fields in chunk, object, and part models from unsigned to signed integers (i64/i16) for consistency and compatibility. Adjusted related service logic to cast values appropriately. Also reduced chunk decoder buffer size from 16 MiB to 4 MiB in object service.

+++ main/service/src/chunk.rs
@@ -62,7 +62,7 @@ impl ChunkMutation {
             id: Set(Uuid::new_v4()),
             object_id: Set(object_id),
             part_id: Set(part_id),
-            index: Set(index),
+            index: Set(index as i64),
             data: Set(data),
             ..Default::default()
         };
